测试接口 查看待处理工单列表、查看工单详情、接单、维修成功，新增并测试了接口 支付工单

+++ BakcEnd/CrispMaintenanceCommon/src/main/java/cn/crisp/entity/Indent.java
@@ -23,7 +23,6 @@ import java.time.LocalDateTime;
 public class Indent implements Serializable {
     private static final long serialVersionUID = 442353225233L;
 
-
     @TableId(value = "id")
     private Long id;
 
@@ -39,9 +38,9 @@ public class Indent implements Serializable {
 
     private String addressDetail;
 
-    private Double latitude;
+    private BigDecimal latitude;
 
-    private Double longitude;
+    private BigDecimal longitude;
 
     private String problem;
